add task column in module
