+++ src/pages/Home/Home.jsx
@@ -1,62 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
-import TaskItem from './TaskItem/TaskItem'
+import { DragDropContext, Droppable } from 'react-beautiful-dnd';
+import TaskColumn from './TaskColumn/TaskColumn'
 import './Home.scss';
 
-
-const statusDisplayNames = {
-    'TODO': 'TO DO',
-    'IN_PROGRESS': 'IN PROGRESS',
-    'REVIEW': 'REVIEW',
-    'DONE': 'DONE'
-};
-
-const priorityColors = {
-    'BLOCKER': '#2f80fa',
-    'HIGH': '#fc6969',
-    'MEDIUM': '#ffbd43b8',
-    'LOW': '#63d562a3'
-};
-
-
-
-const TaskColumn = ({ columnId, tasks, provided }) => (
-    <div className={`task-column ${columnId.toLowerCase().replace('_', '-')}`}>
-        <ColumnHeader title={columnId} count={tasks.length} />
-        <div className="task-column__divider" />
-        <TaskList tasks={tasks} provided={provided} />
-    </div>
-);
-
-const ColumnHeader = ({ title, count }) => (
-    <div className="task-column__header">
-        <h3>{statusDisplayNames[title] || title}</h3>
-    </div>
-);
-
-const TaskList = ({ tasks, provided }) => (
-    <div
-        className="task-container"
-        ref={provided.innerRef}
-        {...provided.droppableProps}
-    >
-        {tasks && tasks.filter(task => task).map((task, index) => (
-            <Draggable
-                key={task.id}
-                draggableId={task.id}
-                index={index}
-            >
-                {(provided) => (
-                    <TaskItem task={task} provided={provided} />
-                )}
-            </Draggable>
-        ))}
-        {provided.placeholder}
-    </div>
-);
-
 const Home = () => {
-
     const [tasks, setTasks] = useState({
         'TODO': [],
         'IN_PROGRESS': [],
