+++ src/pages/Home/TaskColumn/TaskColumn.scss
@@ -50,9 +50,3 @@
 .done .task-column__divider {
     background-color: #5030e5b3; 
 }
-
-.task-container {
-    margin-top: 24px;
-    min-height: 100px;
-    padding: 0 10px;
-}
